+++ src/constants/routes.ts
@@ -18,8 +18,8 @@ const ROUTES = {
       profile: '/mypage/host/profile',
       reservation: '/mypage/host/reservation',
       reviews: '/mypage/host/reviews',
-      registerAccommodation: '/mypage/host/register/accommodation',
-      registerRoom: '/mypage/host/register/room',
+      registerAccommodation: '/mypage/host/register/registerAccommodation',
+      registerRoom: '/mypage/host/register/registerRoom',
     },
   },
   search: '/search',
